#331 refactor header names and update count text
